+++ module/Soutenance/src/Soutenance/Rule/PropositionJuryRuleAwareTrait.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Soutenance\Rule;
+
+trait PropositionJuryRuleAwareTrait
+{
+    protected PropositionJuryRule $propositionJuryRule;
+
+    public function setPropositionJuryRule(PropositionJuryRule $propositionJuryRule): void
+    {
+        $this->propositionJuryRule = $propositionJuryRule;
+    }
+}